requirements.txt
git+https://github.com/Joelkb/cinemagoer
pyrogram>=2.0.30
tgcrypto
pymongo[srv]==3.12.3
motor==2.5.1
marshmallow==3.14.1
umongo==3.0.1
requests
bs4
imdbpy==2021.4.18
shortzy
python-dotenv
